+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Made the site a Progressive Web App (PWA)
 - About page content (first section)
+- robots.txt and canonical link
 
 ### Fixed
 
@@ -57,3 +58,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - components folder structure
 - relative to absolute imports with aliases
 - updated to React 18 and Next.js latest
+- changed image formats to .webp